Extract all decision logic into CacheConfiguration after an idea by @christoph-kluge at https://github.com/WyriHaximus/reactphp-http-middleware-response-cache/pull/2#issuecomment-378215665
